Do not delete the listener socket right after creating it.
The listener server socket file needs to be deleted before we bind it
otherwise it's an "address already in use" error. But it was getting
deleted after the bind, causing the file to disappear, and preventing
anyone from connecting to the listener socket since the socket stops
existing.

This is caused by commit 4f5645db1ddf9c976344ec0c1c4875ee821a2d42